Refactor login page component
