+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import Login from "@/sections/Login";
+import Login from "@/sections/login-view";
 
 export default function LoginPage() {
   return <Login />;
